Add backup sysadmins to katuma servers
In case either Dani or Enrico have a "stuck at the dentist" situation
while I'm on holidays.

Jokes aside, we decided with @Matt-Yorkley to let him use our staging
servers from now on to use it as a proof-of-concept for our goal of
granting access for all sysadmins to all servers. This will let us spot
the bottlenecks and areas we will need to work on.