Add Ha-Ri core team devs SSH key to all servers
If there's a production issue in any of the OFN instances any Ha or Ri
dev can put the sysadmin hat on and fix it.

Access to production machines is also needed to investigate production
logs as well as production data when trying to fix high severity issues,
like happened recently.