Upgrade CI to Bionic and Python 3
We better run our CI on a Python supported version. Not sure which sort
of edge cases we might run into with the latest Ansible but an
unmaintained Python version.

This also requires upgrading to Ubuntu 18.04 but AU has been running
it in production for quite some time so we do know it's safe.