Add config settings for Katuma's server migration
I dropped the following from the existing config:

* `swapfile_size`: We migrate to have more RAM (8GB) thus, no need for swap
anymore
* `enable_nginx_logging`: Datadog is not enabled so this config is set
but not used.
* `disable_datadog`: The new server doesn't have it so a `true` won't
take in effect here.

I added the extra `certbot_domains` entry as documented in
https://github.com/openfoodfoundation/ofn-install/wiki/Migrating-a-Production-Server.

As for the entry `alpha.katuma.org`, the old subdomain we redirect to
`app.katuma.org`, we need to decide whether we still maintain it or not.
It's likely that we'll take the chance to drop it.